vissprite_t::x1/x2 are 'short' instead of 'int'

+++ prboom2/src/r_defs.h
@@ -369,7 +369,7 @@ typedef struct vissprite_s
 {
   mobj_t *thing;
   dboolean flip;
-  int x1, x2;
+  short x1, x2;
   fixed_t gx, gy;              // for line side calculation
   fixed_t gz, gzt;             // global bottom / top for silhouette clipping
   fixed_t startfrac;           // horizontal position of x1
